test: extract config suites
