+++ crates/agent-engine/src/config.rs
@@ -1235,1322 +1235,5 @@ fn validate_supported_model(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use std::io::Write;
-    use std::path::Path;
-    use tempfile::TempDir;
-
-    #[test]
-    fn test_config_default() {
-        let config = Config::default();
-        assert_eq!(config.llm_provider, LlmProvider::OpenAiResponses);
-        assert_eq!(
-            config.google_gemini_generate_content_location,
-            "us-central1"
-        );
-        assert_eq!(
-            config.google_gemini_generate_content_model,
-            "gemini-2.0-flash"
-        );
-        assert_eq!(
-            config.openai_responses_base_url,
-            "https://api.openai.com/v1"
-        );
-        assert_eq!(config.openai_responses_model, "gpt-5.4");
-        assert_eq!(
-            config.openai_chat_completions_base_url,
-            "https://api.openai.com/v1"
-        );
-        assert_eq!(config.openai_chat_completions_model, "gpt-5.4");
-        assert_eq!(
-            config.openai_chat_completions_compatible_base_url,
-            "https://api.openai.com/v1"
-        );
-        assert_eq!(
-            config.openai_chat_completions_compatible_model,
-            "qwen3.5-plus"
-        );
-        assert_eq!(config.openrouter_base_url, "https://openrouter.ai/api/v1");
-        assert_eq!(config.openrouter_model, "moonshotai/kimi-k2.6");
-        assert!(config.openrouter_api_key.is_none());
-        assert!(config.openrouter_app_categories.is_empty());
-        assert_eq!(
-            config.anthropic_messages_base_url,
-            "https://api.anthropic.com/v1"
-        );
-        assert_eq!(config.anthropic_messages_model, "claude-3-5-sonnet-latest");
-        assert_eq!(config.llm_request_timeout_secs, 180);
-        assert_eq!(config.tool_timeout_secs, 30);
-        assert_eq!(config.tool_repeat_limit, 4);
-        assert_eq!(config.context_window_tokens, None);
-        assert_eq!(config.compaction_hard_trigger_ratio, None);
-        assert_eq!(config.compaction_soft_trigger_ratio, None);
-        assert!((config.effective_compaction_hard_trigger_ratio() - 0.8).abs() < f32::EPSILON);
-        assert!((config.effective_compaction_soft_trigger_ratio() - 0.72).abs() < f32::EPSILON);
-        assert_eq!(config.effective_context_window_tokens(), 1_050_000);
-        assert_eq!(config.prompt_snapshot_max_chars, 8000);
-        assert!(!config.prompt_snapshot_enabled);
-        assert!(config.max_tool_loops.is_none());
-        assert_eq!(config.streaming_mode, StreamingMode::Auto);
-        assert_eq!(
-            config.partial_stream_recovery_mode,
-            PartialStreamRecoveryMode::ContinueOnce
-        );
-        assert!(config.skill_overrides.is_empty());
-        // Memory config
-        assert!(config.memory.enabled);
-        assert!(config.memory.strict_store);
-        assert!(config.memory.store_dir.is_none());
-        assert!(!config.durability.required);
-    }
-
-    #[test]
-    fn definition_overlay_content_does_not_require_a_host_file() {
-        let config = Config::default()
-            .with_definition_overlay_content(
-                "tool_repeat_limit = 7\n",
-                Path::new("/lib/pkg/example/agents/root/agent.toml"),
-            )
-            .unwrap();
-
-        assert_eq!(config.tool_repeat_limit, 7);
-    }
-
-    #[test]
-    fn explicit_config_override_must_be_an_existing_absolute_file() {
-        let missing = std::env::temp_dir().join(format!(
-            "alan-missing-config-{}.toml",
-            uuid::Uuid::new_v4().simple()
-        ));
-        let error = Config::load_from_override(Some(missing)).unwrap_err();
-        assert!(
-            error
-                .to_string()
-                .contains("failed to read configuration file"),
-            "unexpected missing-file error: {error:#}"
-        );
-
-        let error = Config::load_from_override(Some("agent.toml".into())).unwrap_err();
-        assert!(
-            error
-                .to_string()
-                .contains("ALAN_CONFIG_PATH must be an absolute path"),
-            "unexpected relative-path error: {error:#}"
-        );
-    }
-
-    #[test]
-    fn test_config_for_google_gemini_generate_content() {
-        let config = Config::for_google_gemini_generate_content(
-            "project",
-            Some("europe-west1"),
-            Some("gemini-2.5-pro"),
-        );
-        assert_eq!(
-            config.llm_provider,
-            LlmProvider::GoogleGeminiGenerateContent
-        );
-        assert_eq!(
-            config.google_gemini_generate_content_project_id,
-            Some("project".to_string())
-        );
-        assert_eq!(
-            config.google_gemini_generate_content_location,
-            "europe-west1"
-        );
-        assert_eq!(
-            config.google_gemini_generate_content_model,
-            "gemini-2.5-pro"
-        );
-        assert!(config.has_google_gemini_generate_content_config());
-        assert!(config.has_llm_config());
-    }
-
-    #[test]
-    fn test_config_for_google_gemini_generate_content_defaults() {
-        let config = Config::for_google_gemini_generate_content("project", None, None);
-        assert_eq!(
-            config.google_gemini_generate_content_location,
-            "us-central1"
-        );
-        assert_eq!(
-            config.google_gemini_generate_content_model,
-            "gemini-2.0-flash"
-        );
-    }
-
-    #[test]
-    fn test_config_for_openai_responses() {
-        let config = Config::for_openai_responses(
-            "sk-test",
-            Some("https://api.openai.com/v1"),
-            Some("gpt-5.4"),
-        );
-        assert_eq!(config.llm_provider, LlmProvider::OpenAiResponses);
-        assert_eq!(config.openai_responses_api_key, Some("sk-test".to_string()));
-        assert_eq!(config.openai_responses_model, "gpt-5.4");
-        assert!(config.has_openai_responses_config());
-        assert!(config.has_llm_config());
-    }
-
-    #[test]
-    fn test_config_for_openai_responses_defaults() {
-        let config = Config::for_openai_responses("sk-test", None, None);
-        assert_eq!(
-            config.openai_responses_base_url,
-            "https://api.openai.com/v1"
-        );
-        assert_eq!(config.openai_responses_model, "gpt-5.4");
-    }
-
-    #[test]
-    fn test_config_for_chatgpt() {
-        let config = Config::for_chatgpt(
-            Some("https://chatgpt.com/backend-api/codex"),
-            Some("gpt-5.3-codex"),
-        );
-        assert_eq!(config.llm_provider, LlmProvider::Chatgpt);
-        assert_eq!(
-            config.chatgpt_base_url,
-            "https://chatgpt.com/backend-api/codex"
-        );
-        assert_eq!(config.chatgpt_model, "gpt-5.3-codex");
-        assert!(config.has_llm_config());
-    }
-
-    #[test]
-    fn test_config_for_chatgpt_defaults() {
-        let config = Config::for_chatgpt(None, None);
-        assert_eq!(
-            config.chatgpt_base_url,
-            "https://chatgpt.com/backend-api/codex"
-        );
-        assert_eq!(config.chatgpt_model, "gpt-5.3-codex");
-    }
-
-    #[test]
-    fn test_config_for_openai_chat_completions() {
-        let config = Config::for_openai_chat_completions(
-            "sk-test",
-            Some("https://api.openai.com/v1"),
-            Some("gpt-5.4"),
-        );
-        assert_eq!(config.llm_provider, LlmProvider::OpenAiChatCompletions);
-        assert_eq!(
-            config.openai_chat_completions_api_key,
-            Some("sk-test".to_string())
-        );
-        assert_eq!(config.openai_chat_completions_model, "gpt-5.4");
-        assert!(config.has_openai_chat_completions_config());
-        assert!(config.has_llm_config());
-    }
-
-    #[test]
-    fn test_config_for_openai_chat_completions_defaults() {
-        let config = Config::for_openai_chat_completions("sk-test", None, None);
-        assert_eq!(
-            config.openai_chat_completions_base_url,
-            "https://api.openai.com/v1"
-        );
-        assert_eq!(config.openai_chat_completions_model, "gpt-5.4");
-    }
-
-    #[test]
-    fn test_config_for_openai_chat_completions_compatible() {
-        let config = Config::for_openai_chat_completions_compatible(
-            "sk-test",
-            Some("https://api.openai.com/v1"),
-            Some("qwen3.5-plus"),
-        );
-        assert_eq!(
-            config.llm_provider,
-            LlmProvider::OpenAiChatCompletionsCompatible
-        );
-        assert_eq!(
-            config.openai_chat_completions_compatible_api_key,
-            Some("sk-test".to_string())
-        );
-        assert_eq!(
-            config.openai_chat_completions_compatible_model,
-            "qwen3.5-plus"
-        );
-        assert!(config.has_openai_chat_completions_compatible_config());
-        assert!(config.has_llm_config());
-    }
-
-    #[test]
-    fn test_config_for_openai_chat_completions_compatible_defaults() {
-        let config = Config::for_openai_chat_completions_compatible("sk-test", None, None);
-        assert_eq!(
-            config.openai_chat_completions_compatible_base_url,
-            "https://api.openai.com/v1"
-        );
-        assert_eq!(
-            config.openai_chat_completions_compatible_model,
-            "qwen3.5-plus"
-        );
-    }
-
-    #[test]
-    fn test_config_for_openrouter() {
-        let config = Config::for_openrouter(
-            "sk-or",
-            None,
-            Some("anthropic/claude-sonnet-4"),
-            Some("https://alan.local"),
-            Some("alan"),
-            vec!["cli-agent".to_string()],
-        );
-        assert_eq!(config.llm_provider, LlmProvider::OpenRouter);
-        assert_eq!(config.openrouter_api_key.as_deref(), Some("sk-or"));
-        assert_eq!(config.openrouter_base_url, "https://openrouter.ai/api/v1");
-        assert_eq!(config.openrouter_model, "anthropic/claude-sonnet-4");
-        assert_eq!(
-            config.openrouter_http_referer.as_deref(),
-            Some("https://alan.local")
-        );
-        assert_eq!(config.openrouter_x_title.as_deref(), Some("alan"));
-        assert_eq!(config.openrouter_app_categories, vec!["cli-agent"]);
-        assert!(config.has_openrouter_config());
-        assert!(config.has_llm_config());
-    }
-
-    #[test]
-    fn test_config_for_anthropic_messages() {
-        let config = Config::for_anthropic_messages(
-            "ak-test",
-            Some("https://api.anthropic.com/v1"),
-            Some("claude-sonnet-4-5"),
-        );
-        assert_eq!(config.llm_provider, LlmProvider::AnthropicMessages);
-        assert_eq!(
-            config.anthropic_messages_api_key,
-            Some("ak-test".to_string())
-        );
-        assert_eq!(config.anthropic_messages_model, "claude-sonnet-4-5");
-        assert!(config.has_anthropic_messages_config());
-        assert!(config.has_llm_config());
-    }
-
-    #[test]
-    fn test_config_for_anthropic_messages_with_options() {
-        let config = Config {
-            llm_provider: LlmProvider::AnthropicMessages,
-            anthropic_messages_api_key: Some("key".to_string()),
-            anthropic_messages_base_url: "https://api.anthropic.com/v1".to_string(),
-            anthropic_messages_model: "claude-3".to_string(),
-            anthropic_messages_client_name: Some("test-client".to_string()),
-            anthropic_messages_user_agent: Some("test-agent/1.0".to_string()),
-            ..Config::default()
-        };
-        assert_eq!(
-            config.anthropic_messages_client_name,
-            Some("test-client".to_string())
-        );
-        assert_eq!(
-            config.anthropic_messages_user_agent,
-            Some("test-agent/1.0".to_string())
-        );
-    }
-
-    #[test]
-    fn test_config_for_anthropic_messages_defaults() {
-        let config = Config::for_anthropic_messages("ak-test", None, None);
-        assert_eq!(
-            config.anthropic_messages_base_url,
-            "https://api.anthropic.com/v1"
-        );
-        assert_eq!(config.anthropic_messages_model, "claude-3-5-sonnet-latest");
-    }
-
-    #[test]
-    fn test_effective_model() {
-        let gemini =
-            Config::for_google_gemini_generate_content("project", None, Some("gemini-2.5-pro"));
-        assert_eq!(gemini.effective_model(), "gemini-2.5-pro");
-
-        let chatgpt = Config::for_chatgpt(None, Some("gpt-5.3-codex"));
-        assert_eq!(chatgpt.effective_model(), "gpt-5.3-codex");
-
-        let openai_responses = Config::for_openai_responses("k", None, Some("gpt-5.4"));
-        assert_eq!(openai_responses.effective_model(), "gpt-5.4");
-
-        let openai_chat_completions =
-            Config::for_openai_chat_completions("k", None, Some("gpt-5.4"));
-        assert_eq!(openai_chat_completions.effective_model(), "gpt-5.4");
-
-        let openai_chat_completions_compatible =
-            Config::for_openai_chat_completions_compatible("k", None, Some("qwen3.5-plus"));
-        assert_eq!(
-            openai_chat_completions_compatible.effective_model(),
-            "qwen3.5-plus"
-        );
-
-        let anthropic = Config::for_anthropic_messages("k", None, Some("claude-3-5-sonnet"));
-        assert_eq!(anthropic.effective_model(), "claude-3-5-sonnet");
-
-        let openrouter = Config::for_openrouter(
-            "sk-or",
-            None,
-            Some("openai/gpt-5.4"),
-            None,
-            None,
-            Vec::new(),
-        );
-        assert_eq!(openrouter.effective_model(), "openai/gpt-5.4");
-    }
-
-    #[test]
-    fn test_has_llm_config_without_api_key() {
-        let mut config = Config {
-            llm_provider: LlmProvider::OpenAiResponses,
-            openai_responses_api_key: None,
-            openai_chat_completions_api_key: None,
-            openai_chat_completions_compatible_api_key: None,
-            ..Config::default()
-        };
-        assert!(!config.has_openai_responses_config());
-        assert!(!config.has_llm_config());
-
-        config.llm_provider = LlmProvider::OpenAiChatCompletions;
-        assert!(!config.has_openai_chat_completions_config());
-        assert!(!config.has_llm_config());
-
-        config.llm_provider = LlmProvider::OpenAiChatCompletionsCompatible;
-        assert!(!config.has_openai_chat_completions_compatible_config());
-        assert!(!config.has_llm_config());
-
-        config.llm_provider = LlmProvider::OpenRouter;
-        config.openrouter_api_key = Some("sk-or".to_string());
-        config.openrouter_model.clear();
-        assert!(!config.has_openrouter_config());
-        assert!(!config.has_llm_config());
-
-        config.llm_provider = LlmProvider::AnthropicMessages;
-        config.anthropic_messages_api_key = None;
-        assert!(!config.has_anthropic_messages_config());
-        assert!(!config.has_llm_config());
-
-        config.llm_provider = LlmProvider::GoogleGeminiGenerateContent;
-        config.google_gemini_generate_content_project_id = None;
-        assert!(!config.has_google_gemini_generate_content_config());
-    }
-
-    #[test]
-    fn test_openai_provider_does_not_treat_compat_key_as_valid_config() {
-        let config = Config {
-            llm_provider: LlmProvider::OpenAiResponses,
-            openai_responses_api_key: None,
-            openai_chat_completions_compatible_api_key: Some("sk-legacy".to_string()),
-            ..Config::default()
-        };
-
-        assert!(!config.has_openai_responses_config());
-        assert!(!config.has_llm_config());
-    }
-
-    #[test]
-    fn test_llm_provider_deserialization() {
-        let gemini: LlmProvider =
-            serde_json::from_str("\"google_gemini_generate_content\"").unwrap();
-        assert_eq!(gemini, LlmProvider::GoogleGeminiGenerateContent);
-
-        let chatgpt: LlmProvider = serde_json::from_str("\"chatgpt\"").unwrap();
-        assert_eq!(chatgpt, LlmProvider::Chatgpt);
-
-        let openai_responses: LlmProvider = serde_json::from_str("\"openai_responses\"").unwrap();
-        assert_eq!(openai_responses, LlmProvider::OpenAiResponses);
-
-        let openai_chat_completions: LlmProvider =
-            serde_json::from_str("\"openai_chat_completions\"").unwrap();
-        assert_eq!(openai_chat_completions, LlmProvider::OpenAiChatCompletions);
-
-        let openai_chat_completions_compatible: LlmProvider =
-            serde_json::from_str("\"openai_chat_completions_compatible\"").unwrap();
-        assert_eq!(
-            openai_chat_completions_compatible,
-            LlmProvider::OpenAiChatCompletionsCompatible
-        );
-
-        let openrouter: LlmProvider = serde_json::from_str("\"openrouter\"").unwrap();
-        assert_eq!(openrouter, LlmProvider::OpenRouter);
-
-        let anthropic: LlmProvider = serde_json::from_str("\"anthropic_messages\"").unwrap();
-        assert_eq!(anthropic, LlmProvider::AnthropicMessages);
-    }
-
-    #[test]
-    fn test_config_from_file() {
-        let temp = TempDir::new().unwrap();
-        let config_path = temp.path().join("test_config.toml");
-
-        let toml_content = r#"
-connection_profile = "openai-main"
-llm_request_timeout_secs = 300
-tool_timeout_secs = 60
-streaming_mode = "off"
-partial_stream_recovery_mode = "off"
-"#;
-
-        let mut file = std::fs::File::create(&config_path).unwrap();
-        file.write_all(toml_content.as_bytes()).unwrap();
-
-        let config = Config::from_file(&config_path).unwrap();
-        assert_eq!(config.connection_profile.as_deref(), Some("openai-main"));
-        assert_eq!(config.llm_request_timeout_secs, 300);
-        assert_eq!(config.tool_timeout_secs, 60);
-        assert_eq!(config.streaming_mode, StreamingMode::Off);
-        assert_eq!(
-            config.partial_stream_recovery_mode,
-            PartialStreamRecoveryMode::Off
-        );
-    }
-
-    #[test]
-    fn test_config_from_file_accepts_model_reasoning_effort() {
-        let temp = TempDir::new().unwrap();
-        let config_path = temp.path().join("test_config.toml");
-        std::fs::write(
-            &config_path,
-            r#"
-model_reasoning_effort = "high"
-"#,
-        )
-        .unwrap();
-
-        let config = Config::from_file(&config_path).unwrap();
-        assert_eq!(config.model_reasoning_effort, Some(ReasoningEffort::High));
-    }
-
-    #[test]
-    fn test_config_from_file_rejects_retired_thinking_budget_as_unknown() {
-        let temp = TempDir::new().unwrap();
-        let config_path = temp.path().join("test_config.toml");
-        std::fs::write(
-            &config_path,
-            r#"
-model_reasoning_effort = "medium"
-thinking_budget_tokens = 2048
-"#,
-        )
-        .unwrap();
-
-        let err = Config::from_file(&config_path).unwrap_err();
-        let message = format!("{err:#}");
-        assert!(message.contains("unknown field"));
-        assert!(message.contains("thinking_budget_tokens"));
-    }
-
-    #[test]
-    fn test_request_control_resolver_uses_model_default_without_budget() {
-        let config = Config::for_openai_responses("sk-test", None, Some("gpt-5.4"));
-        let resolved = crate::resolve_runtime_request_controls(
-            &config,
-            crate::provider_capabilities_for_config(&config),
-            crate::RequestControlIntent::default(),
-        )
-        .unwrap();
-        assert_eq!(resolved.reasoning_effort(), Some(ReasoningEffort::Medium));
-    }
-
-    #[test]
-    fn test_config_from_file_accepts_skill_overrides() {
-        let temp = TempDir::new().unwrap();
-        let config_path = temp.path().join("test_config.toml");
-
-        std::fs::write(
-            &config_path,
-            r#"
-[[skill_overrides]]
-skill = "plan"
-allow_implicit_invocation = false
-"#,
-        )
-        .unwrap();
-
-        let config = Config::from_file(&config_path).unwrap();
-        assert_eq!(config.skill_overrides.len(), 1);
-        assert_eq!(
-            config
-                .skill_overrides
-                .iter()
-                .find(|entry| entry.skill_id == "plan")
-                .unwrap()
-                .allow_implicit_invocation,
-            Some(false)
-        );
-    }
-
-    #[test]
-    fn test_config_from_file_rejects_legacy_skill_override_key() {
-        let temp = TempDir::new().unwrap();
-        let config_path = temp.path().join("legacy-skill-override.toml");
-
-        std::fs::write(
-            &config_path,
-            r#"
-[[skill_overrides]]
-skill_id = "plan"
-allow_implicit_invocation = false
-"#,
-        )
-        .unwrap();
-
-        let err = Config::from_file(&config_path).unwrap_err();
-        let message = format!("{err:#}");
-        assert!(message.contains("failed to parse configuration file"));
-        assert!(message.contains("skill_id"));
-    }
-
-    #[test]
-    fn test_config_from_file_rejects_noncanonical_skill_override_id() {
-        let temp = TempDir::new().unwrap();
-        let config_path = temp.path().join("noncanonical-skill-override.toml");
-
-        std::fs::write(
-            &config_path,
-            r#"
-[[skill_overrides]]
-skill = "repo.review"
-allow_implicit_invocation = false
-"#,
-        )
-        .unwrap();
-
-        let err = Config::from_file(&config_path).unwrap_err();
-        let message = format!("{err:#}");
-        assert!(message.contains("failed to parse configuration file"));
-        assert!(message.contains("repo.review"));
-        assert!(message.contains("repo-review"));
-    }
-
-    #[test]
-    fn test_config_from_file_defaults_skill_overrides_when_omitted() {
-        let temp = TempDir::new().unwrap();
-        let config_path = temp.path().join("test_config.toml");
-
-        std::fs::write(&config_path, "connection_profile = \"openai-main\"\n").unwrap();
-
-        let config = Config::from_file(&config_path).unwrap();
-        assert_eq!(config.connection_profile.as_deref(), Some("openai-main"));
-        assert!(config.skill_overrides.is_empty());
-    }
-
-    #[test]
-    fn test_with_definition_overlays_merges_skill_overrides_field_by_field() {
-        let temp = TempDir::new().unwrap();
-        let overlay_path = temp.path().join("agent.toml");
-        std::fs::write(
-            &overlay_path,
-            r#"
-[[skill_overrides]]
-skill = "plan"
-allow_implicit_invocation = false
-"#,
-        )
-        .unwrap();
-
-        let base = Config {
-            skill_overrides: vec![SkillOverride {
-                skill_id: "plan".to_string(),
-                enabled: Some(false),
-                allow_implicit_invocation: None,
-            }],
-            ..Config::default()
-        };
-        let config = base
-            .with_definition_overlays(std::slice::from_ref(&overlay_path))
-            .unwrap();
-        assert_eq!(
-            config
-                .skill_overrides
-                .iter()
-                .find(|entry| entry.skill_id == "plan")
-                .unwrap()
-                .enabled,
-            Some(false)
-        );
-        assert_eq!(
-            config
-                .skill_overrides
-                .iter()
-                .find(|entry| entry.skill_id == "plan")
-                .unwrap()
-                .allow_implicit_invocation,
-            Some(false)
-        );
-    }
-
-    #[test]
-    fn test_with_definition_overlays_merges_skill_overrides_across_multiple_roots() {
-        let temp = TempDir::new().unwrap();
-        let first_overlay = temp.path().join("base-definition.toml");
-        let second_overlay = temp.path().join("named-definition.toml");
-        std::fs::write(
-            &first_overlay,
-            r#"
-[[skill_overrides]]
-skill = "release-checklist"
-enabled = false
-"#,
-        )
-        .unwrap();
-        std::fs::write(
-            &second_overlay,
-            r#"
-[[skill_overrides]]
-skill = "deploy-checklist"
-allow_implicit_invocation = false
-"#,
-        )
-        .unwrap();
-
-        let config = Config::default()
-            .with_definition_overlays(&[first_overlay, second_overlay])
-            .unwrap();
-
-        assert_eq!(
-            config
-                .skill_overrides
-                .iter()
-                .find(|entry| entry.skill_id == "release-checklist")
-                .unwrap()
-                .enabled,
-            Some(false)
-        );
-        assert_eq!(
-            config
-                .skill_overrides
-                .iter()
-                .find(|entry| entry.skill_id == "deploy-checklist")
-                .unwrap()
-                .allow_implicit_invocation,
-            Some(false)
-        );
-    }
-
-    #[test]
-    fn test_config_from_file_not_found() {
-        let result = Config::from_file(Path::new("/nonexistent/path/config.toml"));
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_config_from_file_invalid_toml() {
-        let temp = TempDir::new().unwrap();
-        let config_path = temp.path().join("invalid.toml");
-
-        std::fs::write(&config_path, "not valid toml {{").unwrap();
-
-        let result = Config::from_file(&config_path);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_builtin_launch_root_agent_configs_parse_as_agent_overlays() {
-        let crate_root = Path::new(env!("CARGO_MANIFEST_DIR"));
-        let paths = [
-            crate_root.join("skills/repo-coding/agents/repo-worker/agent.toml"),
-            crate_root.join("skills/skill-creator/agents/skill-creator/agent.toml"),
-        ];
-
-        for path in paths {
-            Config::from_file(&path)
-                .unwrap_or_else(|err| panic!("failed to parse {}: {err:#}", path.display()));
-        }
-    }
-
-    #[test]
-    fn test_config_from_file_rejects_deprecated_openai_key_names() {
-        let temp = TempDir::new().unwrap();
-        let config_path = temp.path().join("legacy.toml");
-        std::fs::write(
-            &config_path,
-            r#"
-openai_api_key = "sk-test"
-openai_model = "gpt-5"
-"#,
-        )
-        .unwrap();
-
-        let err = Config::from_file(&config_path).unwrap_err();
-        let message = err.to_string();
-        assert!(message.contains("failed to parse configuration file"));
-        assert!(message.contains(&config_path.display().to_string()));
-    }
-
-    #[test]
-    fn test_config_from_file_full() {
-        let temp = TempDir::new().unwrap();
-        let config_path = temp.path().join("full_config.toml");
-
-        let toml_content = r#"
-connection_profile = "anthropic-main"
-llm_request_timeout_secs = 240
-tool_timeout_secs = 45
-max_tool_loops = 10
-tool_repeat_limit = 5
-prompt_snapshot_enabled = true
-prompt_snapshot_max_chars = 10000
-context_window_tokens = 65536
-compaction_hard_trigger_ratio = 0.75
-streaming_mode = "on"
-partial_stream_recovery_mode = "continue_once"
-
-[memory]
-enabled = false
-strict_store = false
-
-[durability]
-required = true
-"#;
-
-        std::fs::write(&config_path, toml_content).unwrap();
-
-        let config = Config::from_file(&config_path).unwrap();
-        assert_eq!(config.connection_profile.as_deref(), Some("anthropic-main"));
-        assert_eq!(config.llm_request_timeout_secs, 240);
-        assert_eq!(config.tool_timeout_secs, 45);
-        assert_eq!(config.max_tool_loops, Some(10));
-        assert_eq!(config.tool_repeat_limit, 5);
-        assert_eq!(config.context_window_tokens, Some(65_536));
-        assert_eq!(config.compaction_hard_trigger_ratio, Some(0.75));
-        assert!((config.effective_compaction_hard_trigger_ratio() - 0.75).abs() < f32::EPSILON);
-        assert!((config.effective_compaction_soft_trigger_ratio() - 0.675).abs() < f32::EPSILON);
-        assert!(config.prompt_snapshot_enabled);
-        assert_eq!(config.prompt_snapshot_max_chars, 10000);
-        assert_eq!(config.streaming_mode, StreamingMode::On);
-        assert_eq!(
-            config.partial_stream_recovery_mode,
-            PartialStreamRecoveryMode::ContinueOnce
-        );
-        // Memory
-        assert!(!config.memory.enabled);
-        assert!(!config.memory.strict_store);
-        assert!(config.durability.required);
-    }
-
-    #[test]
-    fn test_memory_config_default() {
-        let memory = MemoryConfig::default();
-        assert!(memory.enabled);
-        assert!(memory.strict_store);
-        assert!(memory.store_dir.is_none());
-    }
-
-    #[test]
-    fn test_effective_compaction_thresholds_default_soft_from_hard() {
-        let config = Config::default();
-
-        assert!((config.effective_compaction_hard_trigger_ratio() - 0.8).abs() < f32::EPSILON);
-        assert!((config.effective_compaction_soft_trigger_ratio() - 0.72).abs() < f32::EPSILON);
-    }
-
-    #[test]
-    fn test_config_from_file_rejects_retired_compaction_threshold_as_unknown() {
-        let temp_dir = TempDir::new().unwrap();
-        let config_path = temp_dir.path().join("config.toml");
-        std::fs::write(
-            &config_path,
-            r#"
-connection_profile = "openai-main"
-compaction_trigger_ratio = 0.8
-"#,
-        )
-        .unwrap();
-
-        let err = Config::from_file(&config_path).unwrap_err();
-        let message = format!("{err:#}");
-        assert!(message.contains("unknown field"));
-        assert!(message.contains("compaction_trigger_ratio"));
-    }
-
-    #[test]
-    fn test_memory_config_deserialization() {
-        let toml_content = r#"
-enabled = false
-strict_store = false
-store_dir = "/custom/path"
-"#;
-        let memory: MemoryConfig = toml::from_str(toml_content).unwrap();
-        assert!(!memory.enabled);
-        assert!(!memory.strict_store);
-        assert_eq!(memory.store_dir, Some(PathBuf::from("/custom/path")));
-    }
-
-    #[test]
-    fn test_durability_config_deserialization() {
-        let toml_content = r#"
-[durability]
-required = true
-"#;
-        let config: Config = toml::from_str(toml_content).unwrap();
-        assert!(config.durability.required);
-    }
-
-    #[test]
-    fn test_effective_context_window_tokens_uses_explicit_override() {
-        let config = Config {
-            context_window_tokens: Some(42_000),
-            ..Config::default()
-        };
-
-        assert_eq!(config.effective_context_window_tokens(), 42_000);
-    }
-
-    #[test]
-    fn test_effective_context_window_tokens_uses_provider_family_defaults() {
-        let gemini =
-            Config::for_google_gemini_generate_content("project", None, Some("gemini-2.5-pro"));
-        assert_eq!(gemini.effective_context_window_tokens(), 1_048_576);
-
-        let chatgpt = Config::for_chatgpt(None, Some("gpt-5.3-codex"));
-        assert_eq!(chatgpt.effective_context_window_tokens(), 400_000);
-
-        let anthropic =
-            Config::for_anthropic_messages("key", None, Some("claude-3-5-sonnet-latest"));
-        assert_eq!(anthropic.effective_context_window_tokens(), 200_000);
-
-        let openai_responses = Config::for_openai_responses("sk-test", None, Some("gpt-5.4"));
-        assert_eq!(
-            openai_responses.effective_context_window_tokens(),
-            1_050_000
-        );
-
-        let openai_chat_completions =
-            Config::for_openai_chat_completions("sk-test", None, Some("gpt-5.4"));
-        assert_eq!(
-            openai_chat_completions.effective_context_window_tokens(),
-            1_050_000
-        );
-
-        let openai_chat_completions_pro =
-            Config::for_openai_chat_completions("sk-test", None, Some("gpt-5.2-pro"));
-        assert_eq!(
-            openai_chat_completions_pro.effective_context_window_tokens(),
-            400_000
-        );
-
-        let openai_compat = Config::for_openai_chat_completions_compatible(
-            "sk-test",
-            None,
-            Some("bailian/qwen3.5-plus-2026-02-15"),
-        );
-        assert_eq!(openai_compat.effective_context_window_tokens(), 1_000_000);
-
-        let minimax =
-            Config::for_openai_chat_completions_compatible("sk-test", None, Some("MiniMax-M2.5"));
-        assert_eq!(minimax.effective_context_window_tokens(), 204_800);
-
-        let glm = Config::for_openai_chat_completions_compatible("sk-test", None, Some("glm-5"));
-        assert_eq!(glm.effective_context_window_tokens(), 200_000);
-
-        let kimi =
-            Config::for_openai_chat_completions_compatible("sk-test", None, Some("kimi-k2.5"));
-        assert_eq!(kimi.effective_context_window_tokens(), 250_000);
-
-        let deepseek = Config::for_openai_chat_completions_compatible(
-            "sk-test",
-            None,
-            Some("deepseek-reasoner"),
-        );
-        assert_eq!(deepseek.effective_context_window_tokens(), 128_000);
-
-        let unknown = Config::for_openai_chat_completions_compatible(
-            "sk-test",
-            None,
-            Some("my-custom-model"),
-        );
-        assert_eq!(unknown.effective_context_window_tokens(), 32_768);
-    }
-
-    #[test]
-    fn test_with_definition_overlays_merges_model_reasoning_effort() {
-        let temp = TempDir::new().unwrap();
-        let overlay_path = temp.path().join("agent.toml");
-        std::fs::write(&overlay_path, "model_reasoning_effort = \"high\"\n").unwrap();
-
-        let config = Config::for_openai_responses("sk-test", None, Some("gpt-5.4"));
-        let overlaid = config.with_definition_overlays(&[overlay_path]).unwrap();
-
-        assert_eq!(
-            crate::resolve_runtime_request_controls(
-                &overlaid,
-                crate::provider_capabilities_for_config(&overlaid),
-                crate::RequestControlIntent::default(),
-            )
-            .unwrap()
-            .reasoning_effort(),
-            Some(ReasoningEffort::High)
-        );
-    }
-
-    #[test]
-    fn test_with_definition_overlays_preserves_internal_provider_state_for_same_connection_profile()
-    {
-        let temp = TempDir::new().unwrap();
-        let overlay_path = temp.path().join("agent.toml");
-        std::fs::write(&overlay_path, "tool_repeat_limit = 9\n").unwrap();
-
-        let config = Config {
-            connection_profile: Some("openai-main".to_string()),
-            llm_provider: LlmProvider::OpenAiResponses,
-            openai_responses_api_key: Some("sk-test".to_string()),
-            openai_responses_model: "gpt-5.4".to_string(),
-            ..Default::default()
-        };
-
-        let overlaid = config.with_definition_overlays(&[overlay_path]).unwrap();
-
-        assert_eq!(overlaid.connection_profile.as_deref(), Some("openai-main"));
-        assert_eq!(overlaid.llm_provider, LlmProvider::OpenAiResponses);
-        assert_eq!(
-            overlaid.openai_responses_api_key.as_deref(),
-            Some("sk-test")
-        );
-        assert_eq!(overlaid.openai_responses_model, "gpt-5.4");
-        assert_eq!(overlaid.tool_repeat_limit, 9);
-    }
-
-    #[test]
-    fn test_to_provider_config_gemini() {
-        let config = Config::for_google_gemini_generate_content(
-            "my-project",
-            None,
-            Some("gemini-2.0-flash"),
-        );
-        let provider_config = config.to_provider_config().unwrap();
-        // Verify it creates the right config type
-        assert_eq!(
-            provider_config.provider_type,
-            alan_llm::factory::ProviderType::GoogleGeminiGenerateContent
-        );
-        assert_eq!(provider_config.project_id, Some("my-project".to_string()));
-        assert_eq!(provider_config.model, "gemini-2.0-flash");
-    }
-
-    #[test]
-    fn test_to_provider_config_google_gemini_generate_content_missing_project() {
-        let config = Config {
-            llm_provider: LlmProvider::GoogleGeminiGenerateContent,
-            google_gemini_generate_content_project_id: None,
-            ..Config::default()
-        };
-        let result = config.to_provider_config();
-        assert!(result.is_err());
-        assert!(
-            result
-                .unwrap_err()
-                .to_string()
-                .contains("google_gemini_generate_content_project_id")
-        );
-    }
-
-    #[test]
-    fn test_to_provider_config_openai_responses() {
-        let config = Config::for_openai_responses("sk-test", None, Some("gpt-5.4"));
-        let provider_config = config.to_provider_config().unwrap();
-        assert_eq!(
-            provider_config.provider_type,
-            alan_llm::factory::ProviderType::OpenAiResponses
-        );
-        assert_eq!(provider_config.api_key, Some("sk-test".to_string()));
-        assert_eq!(provider_config.model, "gpt-5.4");
-    }
-
-    #[test]
-    fn test_to_provider_config_chatgpt() {
-        let config = Config::for_chatgpt(
-            Some("https://chatgpt.com/backend-api/codex"),
-            Some("gpt-5.3-codex"),
-        );
-        let provider_config = config.to_provider_config().unwrap();
-        assert_eq!(
-            provider_config.provider_type,
-            alan_llm::factory::ProviderType::ChatgptResponses
-        );
-        assert_eq!(provider_config.api_key, None);
-        assert_eq!(
-            provider_config.base_url,
-            Some("https://chatgpt.com/backend-api/codex".to_string())
-        );
-        assert_eq!(provider_config.model, "gpt-5.3-codex");
-        assert_eq!(provider_config.expected_account_id, None);
-    }
-
-    #[test]
-    fn test_to_provider_config_chatgpt_with_account_binding() {
-        let mut config = Config::for_chatgpt(
-            Some("https://chatgpt.com/backend-api/codex"),
-            Some("gpt-5.3-codex"),
-        );
-        config.chatgpt_account_id = Some("acct_123".to_string());
-        let provider_config = config.to_provider_config().unwrap();
-        assert_eq!(
-            provider_config.expected_account_id.as_deref(),
-            Some("acct_123")
-        );
-    }
-
-    #[test]
-    fn test_to_provider_config_openai_responses_missing_key() {
-        let config = Config {
-            llm_provider: LlmProvider::OpenAiResponses,
-            openai_responses_api_key: None,
-            openai_chat_completions_api_key: None,
-            openai_chat_completions_compatible_api_key: None,
-            ..Config::default()
-        };
-        let result = config.to_provider_config();
-        assert!(result.is_err());
-        assert!(
-            result
-                .unwrap_err()
-                .to_string()
-                .contains("openai_responses_api_key")
-        );
-    }
-
-    #[test]
-    fn test_to_provider_config_openai_chat_completions() {
-        let config = Config::for_openai_chat_completions("sk-test", None, Some("gpt-5.4"));
-        let provider_config = config.to_provider_config().unwrap();
-        assert_eq!(
-            provider_config.provider_type,
-            alan_llm::factory::ProviderType::OpenAiChatCompletions
-        );
-        assert_eq!(provider_config.api_key, Some("sk-test".to_string()));
-        assert_eq!(provider_config.model, "gpt-5.4");
-    }
-
-    #[test]
-    fn test_to_provider_config_openai_chat_completions_compatible() {
-        let config =
-            Config::for_openai_chat_completions_compatible("sk-test", None, Some("qwen3.5-plus"));
-        let provider_config = config.to_provider_config().unwrap();
-        assert_eq!(
-            provider_config.provider_type,
-            alan_llm::factory::ProviderType::OpenAiChatCompletionsCompatible
-        );
-        assert_eq!(provider_config.api_key, Some("sk-test".to_string()));
-        assert_eq!(provider_config.model, "qwen3.5-plus");
-    }
-
-    #[test]
-    fn test_to_provider_config_openrouter() {
-        let config = Config::for_openrouter(
-            "sk-or",
-            Some("https://openrouter.example/api/v1"),
-            Some("anthropic/claude-sonnet-4"),
-            Some("https://alan.local"),
-            Some("alan"),
-            vec!["cli-agent".to_string()],
-        );
-        let provider_config = config.to_provider_config().unwrap();
-        assert_eq!(
-            provider_config.provider_type,
-            alan_llm::factory::ProviderType::OpenRouter
-        );
-        assert_eq!(provider_config.api_key.as_deref(), Some("sk-or"));
-        assert_eq!(
-            provider_config.base_url.as_deref(),
-            Some("https://openrouter.example/api/v1")
-        );
-        assert_eq!(provider_config.model, "anthropic/claude-sonnet-4");
-        assert_eq!(
-            provider_config.http_referer.as_deref(),
-            Some("https://alan.local")
-        );
-        assert_eq!(provider_config.x_title.as_deref(), Some("alan"));
-        assert_eq!(
-            provider_config.app_categories,
-            Some(vec!["cli-agent".to_string()])
-        );
-    }
-
-    #[test]
-    fn test_to_provider_config_openrouter_missing_model() {
-        let config = Config {
-            llm_provider: LlmProvider::OpenRouter,
-            openrouter_api_key: Some("sk-or".to_string()),
-            openrouter_model: String::new(),
-            ..Config::default()
-        };
-        let result = config.to_provider_config();
-        assert!(result.is_err());
-        assert!(result.unwrap_err().to_string().contains("openrouter_model"));
-    }
-
-    #[test]
-    fn test_to_provider_config_openai_chat_completions_compatible_accepts_snapshot_and_vendor_prefix()
-     {
-        let config = Config::for_openai_chat_completions_compatible(
-            "sk-test",
-            None,
-            Some("bailian/qwen3.5-plus-2026-02-15"),
-        );
-        let provider_config = config.to_provider_config().unwrap();
-        assert_eq!(
-            provider_config.provider_type,
-            alan_llm::factory::ProviderType::OpenAiChatCompletionsCompatible
-        );
-        assert_eq!(provider_config.model, "bailian/qwen3.5-plus-2026-02-15");
-    }
-
-    #[test]
-    fn test_to_provider_config_openai_chat_completions_compatible_rejects_non_snapshot_variant_suffix()
-     {
-        let config = Config::for_openai_chat_completions_compatible(
-            "sk-test",
-            None,
-            Some("kimi-k2.5-thinking"),
-        );
-        let result = config.to_provider_config();
-        assert!(result.is_err());
-        assert!(result.unwrap_err().to_string().contains("curated catalog"));
-    }
-
-    #[test]
-    fn test_to_provider_config_openai_does_not_fall_back_to_compat_settings() {
-        let config = Config {
-            llm_provider: LlmProvider::OpenAiResponses,
-            openai_responses_api_key: None,
-            openai_chat_completions_compatible_api_key: Some("sk-legacy".to_string()),
-            openai_chat_completions_compatible_base_url: "https://proxy.example/v1".to_string(),
-            openai_chat_completions_compatible_model: "qwen3.5-plus".to_string(),
-            ..Config::default()
-        };
-
-        let result = config.to_provider_config();
-        assert!(result.is_err());
-        assert!(
-            result
-                .unwrap_err()
-                .to_string()
-                .contains("openai_responses_api_key")
-        );
-    }
-
-    #[test]
-    fn test_to_provider_config_openai_rejects_unsupported_model() {
-        let config = Config::for_openai_responses("sk-test", None, Some("gpt-4o"));
-        let result = config.to_provider_config();
-        assert!(result.is_err());
-        assert!(result.unwrap_err().to_string().contains("curated catalog"));
-    }
-
-    #[test]
-    fn test_to_provider_config_openai_chat_completions_compatible_rejects_outdated_model_family() {
-        let config =
-            Config::for_openai_chat_completions_compatible("sk-test", None, Some("kimi-k2"));
-        let result = config.to_provider_config();
-        assert!(result.is_err());
-        assert!(result.unwrap_err().to_string().contains("curated catalog"));
-    }
-
-    #[test]
-    fn test_to_provider_config_anthropic_messages() {
-        let config = Config::for_anthropic_messages("sk-test", None, Some("claude-3"));
-        let provider_config = config.to_provider_config().unwrap();
-        assert_eq!(
-            provider_config.provider_type,
-            alan_llm::factory::ProviderType::AnthropicMessages
-        );
-        assert_eq!(provider_config.api_key, Some("sk-test".to_string()));
-        assert_eq!(provider_config.model, "claude-3");
-    }
-
-    #[test]
-    fn test_to_provider_config_anthropic_messages_with_options() {
-        let config = Config {
-            llm_provider: LlmProvider::AnthropicMessages,
-            anthropic_messages_api_key: Some("key".to_string()),
-            anthropic_messages_base_url: "https://custom.api.com".to_string(),
-            anthropic_messages_model: "claude-3".to_string(),
-            anthropic_messages_client_name: Some("test-client".to_string()),
-            anthropic_messages_user_agent: Some("test-agent/1.0".to_string()),
-            ..Config::default()
-        };
-        let provider_config = config.to_provider_config().unwrap();
-        assert_eq!(
-            provider_config.base_url,
-            Some("https://custom.api.com".to_string())
-        );
-        assert_eq!(provider_config.client_name, Some("test-client".to_string()));
-        assert_eq!(
-            provider_config.user_agent,
-            Some("test-agent/1.0".to_string())
-        );
-    }
-
-    #[test]
-    fn test_to_provider_config_anthropic_messages_missing_key() {
-        let config = Config {
-            llm_provider: LlmProvider::AnthropicMessages,
-            anthropic_messages_api_key: None,
-            ..Config::default()
-        };
-        let result = config.to_provider_config();
-        assert!(result.is_err());
-        assert!(
-            result
-                .unwrap_err()
-                .to_string()
-                .contains("anthropic_messages_api_key")
-        );
-    }
-
-    #[test]
-    fn test_default_functions() {
-        assert_eq!(default_llm_provider(), LlmProvider::OpenAiResponses);
-        assert_eq!(
-            default_google_gemini_generate_content_location(),
-            "us-central1"
-        );
-        assert_eq!(
-            default_google_gemini_generate_content_model(),
-            "gemini-2.0-flash"
-        );
-        assert_eq!(
-            default_openai_responses_base_url(),
-            "https://api.openai.com/v1"
-        );
-        assert_eq!(default_openai_responses_model(), "gpt-5.4");
-        assert_eq!(
-            default_openai_chat_completions_base_url(),
-            "https://api.openai.com/v1"
-        );
-        assert_eq!(default_openai_chat_completions_model(), "gpt-5.4");
-        assert_eq!(
-            default_openai_chat_completions_compatible_base_url(),
-            "https://api.openai.com/v1"
-        );
-        assert_eq!(
-            default_openai_chat_completions_compatible_model(),
-            "qwen3.5-plus"
-        );
-        assert_eq!(
-            default_openrouter_base_url(),
-            "https://openrouter.ai/api/v1"
-        );
-        assert_eq!(default_openrouter_model(), "moonshotai/kimi-k2.6");
-        assert_eq!(
-            default_anthropic_messages_base_url(),
-            "https://api.anthropic.com/v1"
-        );
-        assert_eq!(
-            default_anthropic_messages_model(),
-            "claude-3-5-sonnet-latest"
-        );
-        assert_eq!(default_llm_timeout_secs(), 180);
-        assert_eq!(default_tool_timeout_secs(), 30);
-        assert_eq!(default_prompt_snapshot_max_chars(), 8000);
-        assert_eq!(default_tool_repeat_limit(), 4);
-        assert_eq!(default_streaming_mode(), StreamingMode::Auto);
-        assert_eq!(
-            default_partial_stream_recovery_mode(),
-            PartialStreamRecoveryMode::ContinueOnce
-        );
-    }
-}
+#[path = "config_tests.rs"]
+mod tests;
