+++ crates/agent-engine/src/config_tests.rs
@@ -1,0 +1,8 @@
+use super::*;
+
+#[path = "config_construction_tests.rs"]
+mod construction;
+#[path = "config_file_tests.rs"]
+mod file;
+#[path = "config_provider_tests.rs"]
+mod provider;
